app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "llama-index>=0.13.3",
    "llama-index-embeddings-gemini>=0.4.0",
    "llama-index-llms-gemini>=0.6.0",
    "streamlit>=1.49.1",
]
